Add files via upload
I chose a recycling image to help make the white text stand out more, but I made sure it wasn't too overwhelming. I also added the BHW bottle and a picture of a farmer's market, and I made sure to make those pictures not as bright so that way they flow easily with the color scheme.