Add files via upload
I decided to include a back part of the flyer to give more information about the businesses and restaurants that will be there. I made most of the colors faded to make the text stand out. I also decided to go with a simpler look since the front side of my flyer is busy.