Gimp Final Front Page and Back Page of Flyer
Her is the final draft of the front and back page of my flyer. I've struggled a lot with understanding GIMP, but I'm really appreciative that I now know how to work a photo editing application. Towards the end, I no longer felt the need to watch tutorials for everything and I remembered to engage directly with the individual layers to enhance an image. I think the most important thing to take away from GIMP is to not view an image as a giant picture, but as a collection of objects. Once I viewed my project as a bunch of singular images I was able to enhance each one with drop shadow, contrast, brightness, etc to create one overall flyer that flowed smoothly and looks aesthetically pleasing.